generating base class for start of prblem
