+++ src/mesido/base_problem_mixin.py
@@ -1,0 +1,31 @@
+class BaseProblemMixin:
+    """
+
+    """
+
+    def goal_programming_options(self):
+        """
+        Here we set the goal programming configuration. We use soft constraints for consecutive
+        goals.
+        """
+        options = super().goal_programming_options()
+        options["keep_soft_constraints"] = True
+        return options
+
+    def solver_options(self):
+        """
+        Here we define the solver options. By default we use the open-source solver cbc and casadi
+        solver qpsol.
+        """
+        options = super().solver_options()
+        options["casadi_solver"] = "qpsol"
+        options["solver"] = "highs"
+        return options
+
+    def compiler_options(self):
+        """
+        In this function we set the compiler configuration.
+        """
+        options = super().compiler_options()
+        options["resolve_parameter_values"] = True
+        return options
