+++ src/mesido/electricity_physics_mixin.py
@@ -7,6 +7,7 @@ from typing import List, Tuple
 import casadi as ca
 
 from mesido.base_component_type_mixin import BaseComponentTypeMixin
+from mesido.base_problem_mixin import BaseProblemMixin
 
 import numpy as np
 
@@ -41,7 +42,8 @@ class ElectrolyzerOption(IntEnum):
     LINEARIZED_THREE_LINES_EQUALITY = 3
 
 
-class ElectricityPhysicsMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationProblem):
+class ElectricityPhysicsMixin(BaseProblemMixin, BaseComponentTypeMixin,
+                              CollocatedIntegratedOptimizationProblem):
     """
     This class is used to model the physics of an electricity network with its assets. We model
     the different components with variety of linearization strategies.
@@ -937,30 +939,3 @@ class ElectricityPhysicsMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimi
         constraints.extend(self.__electricity_producer_set_point_constraints(ensemble_member))
 
         return constraints
-
-    def goal_programming_options(self):
-        """
-        Here we set the goal programming configuration. We use soft constraints for consecutive
-        goals.
-        """
-        options = super().goal_programming_options()
-        options["keep_soft_constraints"] = True
-        return options
-
-    def solver_options(self):
-        """
-        Here we define the solver options. By default we use the open-source solver cbc and casadi
-        solver qpsol.
-        """
-        options = super().solver_options()
-        options["casadi_solver"] = "qpsol"
-        options["solver"] = "highs"
-        return options
-
-    def compiler_options(self):
-        """
-        In this function we set the compiler configuration.
-        """
-        options = super().compiler_options()
-        options["resolve_parameter_values"] = True
-        return options
